backend/pyproject.toml
[project]
name = "website-design-scorer"
version = "0.1.0"
description = "Website Design Scorer Backend - AI-powered design analysis and reporting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    # Phase 2: Screenshot & Storage
    "playwright>=1.40.0",
    "cloudinary>=1.37.0",
    "pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "aiofiles>=23.0.0",
    # Phase 3: AI Analysis Engine
    "cssutils>=2.10.0",
    "tinycss2>=1.2.0",
    "colorthief>=0.2.1",
    "webcolors>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
